requirements.txt
# Core dependencies
jericho
python-dotenv

# LLM providers
huggingface_hub
openai
anthropic

# MCP Server
fastmcp

# Function calling (optional, for the alternative approach)
langchain-core
